requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0
opencv-python-headless==4.9.0.80
tensorflow==2.15.0
openai==1.12.0
plotly==5.18.0
scikit-learn==1.4.0
requests==2.31.0
python-dotenv==1.0.1
streamlit-chat==0.1.1
streamlit-option-menu==0.3.12
google-generativeai==0.3.2 